(new) wishlist manager

+++ Application/IPriceManager.cs
@@ -2,5 +2,5 @@
 
 public interface IPriceManager
 {
-    public Task<(string? name, string finalPrice, int discount)> GetSteamPriceAsync(int gameId, string userRegion);
+    public Task<(string? name, string finalPrice, int discount)> GetSteamPriceAsync(long gameId, string userRegion);
 }
